back: Split the two files this branch pushed past the guide's line.
`docs/tech/development-guide.md` has one convention with a numeric trigger —
"File > 800 lines | Split now" — and this branch fires it twice.
`reviewQueueController.ts` went from 792 lines to 867 when the `contents` card
learned to list its rows, and `locationWriter.ts` from 911 to 1076 when the
per-point diff landed. Both are the row below it as well: a new responsibility
added to an already-large file, which the guide says to extract from the start.

Four modules, each drawn on a seam the branch itself had already exposed:

- `reviewQueueContents.ts` — the `contents` and `withdrawn` queries. They read
  `experience_locations` rather than `experiences`, they are about a component
  rather than an object, and between them they carry the only per-row lists the
  queue returns — so the cap that bounds those lists lives with them.
- `reviewQueueContext.ts` — what a card shows about the object it asks about.
  Two files build queue rows now, and a shared fragment is the only arrangement
  in which the eight cards cannot drift into showing different amounts.
- `locationPairing.ts` — what the writer's statements are built *from*: identity
  (`samePointSql`, `claimedPointSql`), the guard that keeps a claimed column, and
  the reading of what a kept row's own columns say happened to it.
  `contentsChangeSet.ts` already consumed the last of those from outside the
  file, which is the seam this follows. The header's identity measurements went
  with the fragments they justify.
- `locationIncoming.ts` — the source's list before anything is known about the
  store: its CTE, its parameters, and the duplicates the source itself ships.

Behaviour is unchanged and the moved text is unchanged; every assertion that
reads these statements still passes, including the ones that match the SQL
itself. `reviewQueueController.ts` is 514 lines, `locationWriter.ts` 798.
